requirements.txt
streamlit==1.24.0
streamlit_chat==0.1.1
bardapi==0.1.24
Pillow<10.0.0

